wip: factoryクラスを分離した
<body: why>

---
add: (機能・ファイルを)追加する

fix: (コードを)修正する

clean: (コードを)改善する; 名前の変更/コードの整形

update: (パッケージやドキュメントなどを更新する)

remove: (ファイルやコードを)除去する

rename: (ファイル名を)変更する

change: (AをBに)変更する

ref: https://www.tam-tam.co.jp/tipsnote/program/post16686.html
